fix: move Expandable in ui directory Signed-off-by: Philippe Martin <phmartin@redhat.com>
